# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/recipes": {
            "get": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Get all recipes",
                "operationId": "d4042afb2fbbddc6df47e660bab9cea7",
                "responses": {
                    "200": {
                        "description": "A list of recipes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Recipe"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Create a new recipe",
                "operationId": "b8e5dc90090b088f52bf6ba8ada1a43f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Spaghetti Bolognese"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "A classic Italian pasta dish"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "image_url"
                                    },
                                    "video": {
                                        "type": "string",
                                        "example": "video_url"
                                    },
                                    "total_time": {
                                        "type": "integer",
                                        "example": 45
                                    },
                                    "preparation_time": {
                                        "type": "integer",
                                        "example": 15
                                    },
                                    "rest_time": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "cooking_time": {
                                        "type": "integer",
                                        "example": 20
                                    },
                                    "title_reference": {
                                        "type": "string",
                                        "example": "Reference Title"
                                    },
                                    "episode_reference": {
                                        "type": "string",
                                        "example": "Episode 1"
                                    },
                                    "description_reference": {
                                        "type": "string",
                                        "example": "Description of the reference"
                                    },
                                    "logo_platform_reference": {
                                        "type": "string",
                                        "example": "https://example.com/logo.png"
                                    },
                                    "logo_platform_url_reference": {
                                        "type": "string",
                                        "example": "https://example.com"
                                    },
                                    "image_recipe_reference": {
                                        "type": "string",
                                        "example": "https://example.com/image.jpg"
                                    },
                                    "ingredients": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "string",
                                                    "example": "2"
                                                },
                                                "unit": {
                                                    "type": "string",
                                                    "example": "pieces"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "diets": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "categories": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Recipe created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Recipe"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "accessToken": []
                    }
                ]
            }
        },
...
        "/recipes/filter": {
            "get": {
                "tags": [
                    "Recipes"
                ],
                "summary": "Filter recipes",
                "operationId": "5d394e75d913fe6cc6119b6f55e6a478",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Filter by recipe title",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Filter by category ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Filtered list of recipes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Recipe"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Category": {
                "title": "Category",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Entrée"
                    },
                    "image": {
                        "type": "string",
                        "example": "https://img-3.journaldesfemmes.fr/KX37xZuExSWZtfhyN7iS-AYL3jU=/1240x/smart/a99c4f96ccfa47cc830cbde4f896eaa8/ccmcms-jdf/39957548.jpg"
                    }
                },
                "type": "object"
            },
            "Diet": {
                "title": "Diet",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Vegetarian"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-09-27T18:41:01.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-09-27T18:41:01.000000Z"
                    }
                },
                "type": "object"
            },
            "Ingredient": {
                "title": "Ingredient",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Pomme"
                    },
                    "quantity": {
                        "type": "string",
                        "example": "2"
                    },
                    "unit": {
                        "type": "string",
                        "example": "pieces"
                    }
                },
                "type": "object"
            },
            "Recipe": {
                "title": "Recipe",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Spaghetti Bolognese"
                    },
                    "description": {
                        "type": "string",
                        "example": "A classic Italian pasta dish"
                    },
                    "image": {
                        "type": "string",
                        "example": "image_url"
                    },
                    "video": {
                        "type": "string",
                        "example": "video_url"
                    },
                    "total_time": {
                        "type": "integer",
                        "example": 45
                    },
                    "preparation_time": {
                        "type": "integer",
                        "example": 15
                    },
                    "rest_time": {
                        "type": "integer",
                        "example": 10
                    },
                    "cooking_time": {
                        "type": "integer",
                        "example": 20
                    },
                    "title_reference": {
                        "type": "string",
                        "example": "Reference Title"
                    },
                    "episode_reference": {
                        "type": "string",
                        "example": "Episode 1"
                    },
                    "diets": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Diet"
                        }
                    },
                    "ingredients": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Ingredient"
                        }
                    },
                    "categories": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-09-27T18:41:01.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-09-27T18:41:01.000000Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "accessToken": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}